修改 music.php

+++ music.php
@@ -1,7 +1,8 @@
 <?php
 require 'r/fun.php';
 require 'x/mysql.class.php';
-$kv->delete('music_xiami_list');
+if(isset($_GET['m']) && $_GET['m'] = 'del')
+	$kv->delete('music_xiami_list');
 $r = $kv->get('music_xiami_list');
 if(!$r){
 	$arr = array();
